Updated Region.py
slightly changed constructor to accept numerical values as strings and convert to floats during assignments (instead of converting to floats before providing the parameter). Changed the syntax of the getCoordinates() function from tuple(a, b) to (a, b) so that the values don't need to be converted to a list and then a tuple. Debugged using the unit tests.