Problem in Country.py
Added a global pointer to the current file position so that you can initialize multiple Country objects. For example, without this, the file "data.txt" (which contains all our data) would be reopened at the beginning of the file for each new object, resulting in all the objects being exactly the same. Using the pointer we can open the file at whatever stage we left off at and continue from there. I think this is a much better solution than storing the data for each country in a separate file. Let me know if you think of any other solutions that might work out well.