GUI Home Screen Code (Version 1)
Added in all needed elements for the homescreen code. Specific file names/paths will need to be modified to call from GitHub instead of the laptop. Additionally, only the Home and Map buttons display in this state, as the remaining widgets need to be transistioned from the grid geometry manager in Tkinter to the place geometry manager. This is because the grid manager does not allow for dynamic scaling, and was not able to place objects in a way that met out initial designs for the home screen.