Suggestions for filepaths
just to reiterate. The use of '.' goes up one level and the use of '/' goes down one level. This applies to the **current** directory i.e. any file in the Air_Pollution_App_CS directory. You do not have to go all the way up to C. To begin with, how can you be sure that everybody has it saved the same way as you? What if somebody has it saved in C:/users/name/program files/useless apps/misc apps/filler folder/spam folder/Air-Pollution-App-CS/. All that you can be certain of is the directory of the actual project, not the client's computer.