Cleaned up GUI, modified to work in VS Code
After finishing the application up, we decided to use some of the remaining time before the due date to clean up the GUI. However, some of our team members ran into issues, as they were using a different IDE than the GUI was created in (VS Code, the GUI was made in PyCharm initially). While they were able to install PyCharm and work on the program, it did bring up the fact that the program may not properly run on all devices. Thus, we started to tweak the program to make it run properly on as many IDEs as possible.